method updateCartCount set as export to use it in different pages. ProductDetails double quote fixed. Item counter icon style edited
