+++ src/css/style.css
@@ -135,9 +135,9 @@ button {
 
 .cart-count-badge {
   position: absolute;
-  top: -5px;
-  right: -10px;
-  background-color: red;
+  top: 5px;
+  right: -8px;
+  background-color: #f0a868;
   color: white;
   font-size: 0.75rem;
   padding: 2px 6px;
