+++ src/js/product.js
@@ -2,6 +2,11 @@ import { getParam } from "./utils.mjs";
 import ProductData from "./ProductData.mjs";
 import ProductDetails from "./ProductDetails.mjs";
 
+import { updateCartCount } from "../js/utils.mjs";
+
+document.addEventListener("DOMContentLoaded", updateCartCount);
+
+
 const dataSource = new ProductData("tents");
 const productId = getParam("product");
